# HR-forge-ui/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Vhark",
    "description": "OpenApi documentation for Hr-forge-api",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Vhark",
      "email": "contact@vhark.com"
    },
    "license": {
      "name": "No license"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local ENV"
    },
    {
      "url": "https://some-url.com",
      "description": "PROD ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/position/{id-or-name}": {
      "get": {
        "tags": [
          "Position"
        ],
        "summary": "Find position by ID or name",
        "description": "Endpoint for retrieving a position by ID or name.",
        "operationId": "getPosition",
        "parameters": [
          {
            "name": "id-or-name",
            "in": "path",
            "description": "Position ID or name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Position"
        ],
        "summary": "Update position",
        "description": "Endpoint for updating an existing position by ID or name.",
        "operationId": "updatePosition",
        "parameters": [
          {
            "name": "id-or-name",
            "in": "path",
            "description": "Position ID or name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Position details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Position"
        ],
        "summary": "Delete position",
        "description": "Endpoint for deleting a position by ID or name.",
        "operationId": "deletePosition",
        "parameters": [
          {
            "name": "id-or-name",
            "in": "path",
            "description": "Position ID or name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/position": {
      "get": {
        "tags": [
          "Position"
        ],
        "summary": "Find all positions",
        "description": "Endpoint for retrieving a paginated list of all positions.",
        "operationId": "getAllPositions",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "example": 10
          },
          {
            "name": "sortField",
            "in": "query",
            "description": "Field to sort by",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "name"
          },
          {
            "name": "sortDirection",
            "in": "query",
            "description": "Sort direction (ASC or DESC)",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "ASC"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponsePositionResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Position"
        ],
        "summary": "Create position",
        "description": "Endpoint for creating a new position.",
        "operationId": "createPosition",
        "requestBody": {
          "description": "Position details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/position/search": {
      "get": {
        "tags": [
          "Position"
        ],
        "summary": "Search positions by name",
        "description": "Endpoint for searching positions by a part of their name.",
        "operationId": "searchPositionsByName",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Part of the position name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Engineer"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "example": 10
          },
          {
            "name": "sortField",
            "in": "query",
            "description": "Field to sort by",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "name"
          },
          {
            "name": "sortDirection",
            "in": "query",
            "description": "Sort direction (ASC or DESC)",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "ASC"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponsePositionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PositionRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "PageResponsePositionResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PositionResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "PositionResponse": {
        "type": "object",
        "properties": {
          "positionId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}